pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.8"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.28.0"
stripe = "^11.3.0"
forex-python = "^1.8"
celery = "^5.4.0"
eventlet = "^0.38.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
